examples/DocTalk/app/requirements.txt
beautifulsoup4==4.12.2
datasets==2.16.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
octoai-sdk==0.2.0
pinecone-datasets==0.6.2
pinecone-client
pinecone-client[grpc]
playwright==1.40.0
python-dotenv==1.0.0
tiktoken==0.5.2
transformers==4.36.2